fixes #2254 - Add STI to hosts table
