+++ app/helpers/hosts_helper.rb
@@ -238,10 +238,7 @@ module HostsHelper
         button_group(
             link_to_if_authorized("Run puppet", hash_for_puppetrun_host_path(:id => host).merge(:auth_action => :edit),
                                   :disabled => !Setting[:puppetrun],
-                                  :title => "Trigger a puppetrun on a node; requires that puppet run is enabled"),
-            link_to_if_authorized("All Puppet Classes", hash_for_storeconfig_klasses_host_path(:id => host).merge(:auth_action => :read),
-                                  :disabled => host.resources.count == 0,
-                                  :title => "Show all host puppet classes, requires storeconfigs")
+                                  :title => "Trigger a puppetrun on a node; requires that puppet run is enabled")
         ),
         button_group(
             link_to_if_authorized("Delete", hash_for_host_path(:id => host, :auth_action => :destroy),
