+++ app/models/environment.rb
@@ -1,8 +1,9 @@
 class Environment < ActiveRecord::Base
   include Taxonomix
+
   has_many :environment_classes, :dependent => :destroy
   has_many :puppetclasses, :through => :environment_classes, :uniq => true
-  has_many :hosts
+  has_many_hosts
   has_many :trends, :as => :trendable, :class_name => "ForemanTrend"
 
   validates_presence_of :name
