+++ app/models/host/base.rb
@@ -1,0 +1,8 @@
+require 'facts_importer'
+
+module Host
+  class Base < ActiveRecord::Base
+    include Foreman::STI
+    set_table_name :hosts
+  end
+end
