+++ app/models/host_class.rb
@@ -1,6 +1,6 @@
 class HostClass < ActiveRecord::Base
   audited :associated_with => :host
-  belongs_to :host
+  belongs_to :host, :foreign_key => :host_id
   belongs_to :puppetclass
 
   validates_presence_of :host_id, :puppetclass_id
