+++ app/models/medium.rb
@@ -1,8 +1,9 @@
 class Medium < ActiveRecord::Base
   include Authorization
   include Taxonomix
+
   has_and_belongs_to_many :operatingsystems
-  has_many :hosts
+  has_many_hosts
 
   # We need to include $ in this as $arch, $release, can be in this string
   VALID_NFS_PATH=/^([-\w\d\.]+):(\/[\w\d\/\$\.]+)$/
