+++ app/models/model.rb
@@ -1,6 +1,7 @@
 class Model < ActiveRecord::Base
   include Authorization
-  has_many :hosts
+
+  has_many_hosts
   has_many :trends, :as => :trendable, :class_name => "ForemanTrend"
   before_destroy EnsureNotUsedBy.new(:hosts)
   validates_uniqueness_of :name
