+++ app/models/ptable.rb
@@ -4,7 +4,8 @@
 # modified version of one of these in textual form
 class Ptable < ActiveRecord::Base
   include Authorization
-  has_many :hosts
+
+  has_many_hosts
   has_and_belongs_to_many :operatingsystems
   before_destroy EnsureNotUsedBy.new(:hosts)
   validates_uniqueness_of :name
