+++ app/models/report.rb
@@ -1,7 +1,8 @@
 class Report < ActiveRecord::Base
   include Authorization
   include ReportCommon
-  belongs_to :host
+
+  belongs_to_host
   has_many :messages, :through => :logs
   has_many :sources, :through => :logs
   has_many :logs, :dependent => :destroy
