+++ app/models/subnet.rb
@@ -2,7 +2,8 @@ require 'ipaddr'
 class Subnet < ActiveRecord::Base
   include Authorization
   include Taxonomix
-  has_many :hosts
+
+  has_many_hosts
   belongs_to :dhcp, :class_name => "SmartProxy"
   belongs_to :tftp, :class_name => "SmartProxy"
   belongs_to :dns,  :class_name => "SmartProxy"
