+++ config/initializers/hostable.rb
@@ -1,0 +1,3 @@
+class ActiveRecord::Base
+  extend Host::Hostmix
+end
