+++ test/fixtures/hosts.yml
@@ -1,6 +1,7 @@
 # Read about fixtures at http://ar.rubyonrails.org/classes/Fixtures.html
 
 one:
+  type: Host::Managed
   name: my5name.mydomain.net
   ip: "2.3.4.1"
   mac: aabbCCddeeee
@@ -17,6 +18,7 @@ one:
   organization: organization1
 
 two:
+  type: Host::Managed
   name: temp.yourdomain.net
   ip: "2.3.4.5"
   mac: "aa:bb:cc:dd:ee:ff"
@@ -32,6 +34,7 @@ two:
   location: location1
 
 otherfullhost:
+  type: Host::Managed
   name: otherfullhost
   mac: aabbecddee11
   ip: 2.3.4.6
@@ -45,6 +48,7 @@ otherfullhost:
   compute_resource: one
 
 anotherfullhost:
+  type: Host::Managed
   name: anotherfullhost
   mac: aabbecddee22
   ip: 2.3.4.7
@@ -58,6 +62,7 @@ anotherfullhost:
   compute_resource: one
 
 myfullhost:
+  type: Host::Managed
   name: myfullname.mydomain.net
   ip: 3.3.4.5
   mac: 112233445566
@@ -70,6 +75,7 @@ myfullhost:
   compute_resource: one
 
 redhat:
+  type: Host::Managed
   name: temp01.yourdomain.net
   ip: 2.3.4.100
   mac: ad:bb:cc:dd:ee:ff
@@ -85,6 +91,7 @@ redhat:
   compute_resource: one
 
 ubuntu:
+  type: Host::Managed
   name: temp-01.yourdomain.net
   ip: 2.3.4.105
   mac: ac:bb:cc:dd:ee:ff
@@ -100,6 +107,7 @@ ubuntu:
   compute_resource: one
 
 minimal:
+  type: Host::Managed
   name: temp-02.useless.net
   ip: 2.3.4.16
   mac: ac:bb:cc:aa:ee:ff
@@ -113,6 +121,7 @@ minimal:
   compute_resource: one
 
 sol10host:
+  type: Host::Managed
   name: sol10host
   ip: 2.3.4.9
   mac: 44:55:66:77:88:99
@@ -127,6 +136,7 @@ sol10host:
   compute_resource: one
 
 suse:
+  type: Host::Managed
   name: suse01.yourdomain.net
   ip: 2.3.4.205
   mac: ad:bb:cc:da:ae:ff
@@ -142,6 +152,7 @@ suse:
   compute_resource: one
 
 dhcp:
+  type: Host::Managed
   name: dhcp.mydomain.net
   ip: "2.3.4.110"
   mac: da:bb:cc:dd:ee:ee
@@ -156,6 +167,7 @@ dhcp:
   compute_resource: one
 
 sp_dhcp:
+  type: Host::Managed
   name: sdhcp.mydomain.net
   ip: "2.3.4.99"
   mac: da:bb:cc:dd:ee:ae
