+++ test/test_helper.rb
@@ -10,6 +10,7 @@ class ActiveSupport::TestCase
 
   fixtures :all
 
+  set_fixture_class({ :hosts => Host::Base })
   # Add more helper methods to be used by all tests here...
 
   def logger
